+++ src/adapt/vtree.js
@@ -487,6 +487,12 @@ adapt.vtree.LayoutContext.prototype.applyFootnoteStyle = function(vertical, elem
 adapt.vtree.LayoutContext.prototype.peelOff = function(nodeContext, nodeOffset) {};
 
 /**
+ * Process a block-end edge of a fragmented block.
+ * @param {adapt.vtree.NodeContext} nodeContext
+ */
+adapt.vtree.LayoutContext.prototype.processFragmentedBlockEdge = function(nodeContext) {};
+
+/**
  * Returns if two NodePositions represents the same position in the document.
  * @param {!adapt.vtree.NodePosition} nodePosition1
  * @param {!adapt.vtree.NodePosition} nodePosition2
@@ -832,6 +838,23 @@ adapt.vtree.NodeContext.prototype.getContainingBlockForAbsolute = function() {
 		parent = parent.parent;
 	}
 	return null;
+};
+
+/**
+ * Walk up NodeContext tree (starting from itself) and call the callback for each block,
+ * until a NodeContext which establishes a block formatting context is reached.
+ * @param {!function(!adapt.vtree.NodeContext)} callback
+ */
+adapt.vtree.NodeContext.prototype.walkBlocksUpToBFC = function(callback) {
+	var nodeContext = this;
+	while (nodeContext) {
+		if (!nodeContext.inline) {
+			callback(nodeContext);
+		}
+		if (nodeContext.establishesBFC)
+			break;
+		nodeContext = nodeContext.parent;
+	}
 };
 
 /**